pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "alt-text-llm"
version = "0.1.1"
description = "AI-powered alt text generation and labeling tools for markdown content"
readme = "README.md"
authors = [{name = "TurnTrout"}]
license = "MIT"
requires-python = ">=3.11"
keywords = ["alt-text", "accessibility", "markdown", "llm", "ai"]

dependencies = [
    "gitpython",
    "requests",
    "ruamel.yaml",
    "markdown-it-py",
    "rich",
    "tqdm",
]

[project.optional-dependencies]
dev = ["pytest", "mypy", "types-requests", "pytest-asyncio"]

[tool.pytest.ini_options]
addopts = "--cov=alt_text_llm --cov-report=html -n auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore:rich is experimental/alpha:tqdm.std.TqdmExperimentalWarning",
]

[project.scripts]
alt-text-llm = "alt_text_llm.main:main"

[project.urls]
Repository = "https://github.com/alexander-turner/alt-text-llm"
